I realized one of the most useful features of JDT is not having to write nasty imports... :-)
git-svn-id: http://svn.dsource.org/projects/descent/trunk@1026 6538119e-2221-0410-ba12-9a5a7e5062c7

+++ descent.ui/src/descent/internal/corext/util/IFileTypeInfo.java
@@ -32,7 +32,7 @@ public class IFileTypeInfo extends TypeInfo {
 	private final String fExtension;
 	private final String fPath;
 	
-	public IFileTypeInfo(String pkg, String name, char[][] enclosingTypes, int modifiers, String project, String sourceFolder, String file, String extension, String path) {
+	public IFileTypeInfo(String pkg, String name, char[][] enclosingTypes, long modifiers, String project, String sourceFolder, String file, String extension, String path) {
 		super(pkg, name, enclosingTypes, modifiers);
 		fProject= project;
 		fFolder= sourceFolder;
